+++ src/main/java/rocks/blackblock/screenbuilder/slots/SlotBuilder.java
@@ -11,12 +11,12 @@ import net.minecraft.item.Items;
 import net.minecraft.screen.slot.Slot;
 import net.minecraft.screen.slot.SlotActionType;
 import org.jetbrains.annotations.Nullable;
-import rocks.blackblock.screenbuilder.interfaces.SlotStackChecker;
-import rocks.blackblock.screenbuilder.inventories.EmptyInventory;
 import rocks.blackblock.screenbuilder.TexturedScreenHandler;
 import rocks.blackblock.screenbuilder.interfaces.GuiListener;
 import rocks.blackblock.screenbuilder.interfaces.SlotAccessListener;
+import rocks.blackblock.screenbuilder.interfaces.SlotStackChecker;
 import rocks.blackblock.screenbuilder.interfaces.SlotWrapperListener;
+import rocks.blackblock.screenbuilder.inventories.EmptyInventory;
 
 import java.util.ArrayList;
 import java.util.List;